requirements.txt
Flask==1.1.2
Flask-Cors>=3.0.9
flask-marshmallow==0.12.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
requests
python-dotenv
plyer
pydub